Move logging exporter into an independent package.
