+++ examples/src/main/java/io/opencensus/examples/trace/MultiSpansTracing.java
@@ -16,7 +16,7 @@ package io.opencensus.examples.trace;
 import io.opencensus.trace.Span;
 import io.opencensus.trace.Tracer;
 import io.opencensus.trace.Tracing;
-import io.opencensus.trace.export.SpanExporter.LoggingHandler;
+import io.opencensus.trace.export.LoggingExportHandler;
 
 /** Example showing how to directly create a child {@link Span} and add annotations. */
 public final class MultiSpansTracing {
@@ -39,7 +39,7 @@ public final class MultiSpansTracing {
    * @param args the main arguments.
    */
   public static void main(String[] args) {
-    LoggingHandler.register(Tracing.getExportComponent().getSpanExporter());
+    LoggingExportHandler.register(Tracing.getExportComponent().getSpanExporter());
     doWork();
   }
 }
