+++ exporters/logging/build.gradle
@@ -1,0 +1,10 @@
+description = 'OpenCensus Logging Exporter'
+
+dependencies {
+    compile project(':opencensus-api'),
+            libraries.guava
+
+    testCompile project(':opencensus-api')
+
+    signature "org.codehaus.mojo.signature:java16:+@signature"
+}
